requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-extensions==3.2.3
python-decouple==3.8
drf-yasg==1.21.7
django-model-utils==4.5.1